% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LukaFunctions.R
\name{getArtist_ID}
\alias{getArtist_ID}
\title{Retrieves an artist ID from a general name search query}
\usage{
getArtist_ID(artist = NA, authentication_token = auth_token)
}
\arguments{
\item{artist}{The artist to search for as a string}

\item{authentication_token}{Predefined argument after running get_authentication_token()}
}
\value{
Dataframe containing the artist name, ID, popularity, and related genres
}
\description{
Retrieves an artist ID from a general name search query
}
\examples{
getArtist_ID('alt j')
}
